requirements.txt
requests
kafka-python
pyspark
websocket-client
